+++ lib/apical/plugs/path.ex
@@ -57,8 +57,9 @@ defmodule Apical.Plugs.Path do
 
     conn
     |> Map.update!(:params, &Map.merge(&1, params))
-    |> Parameter.warn_deprecated(params, :path, operations)
-    |> Parameter.validate(params, :path, operations)
+    |> Parameter.warn_deprecated(:path, operations)
+    |> Parameter.custom_marshal(:path, operations)
+    |> Parameter.validate(:path, operations)
   end
 
   @impl Apical.Plugs.Parameter