+++ test/parameters/header_test.exs
@@ -107,6 +107,20 @@ defmodule ApicalTest.Parameters.HeaderTest do
               - name: style-custom-override
                 in: header
                 style: x-custom
+              - name: marshal-defined
+                in: header
+                schema:
+                  oneOf:
+                    - type: integer
+                    - type: boolean
+              - name: validate-disabled
+                in: header
+                schema:
+                  type: boolean
+              - name: marshal-disabled
+                in: header
+                schema:
+                  type: integer
         "/override":
           get:
             operationId: headerParamOverride
@@ -123,7 +137,13 @@ defmodule ApicalTest.Parameters.HeaderTest do
       parameters: [
         "style-custom-override": [
           styles: [{"x-custom", {__MODULE__, :x_custom, ["by parameter"]}}]
-        ]
+        ],
+        "marshal-defined": [
+          # also test `{atom, list}` style here
+          marshal: {:defined_marshalling, [:atom]}
+        ],
+        "validate-disabled": [validate: false],
+        "marshal-disabled": [marshal: false]
       ],
       operation_ids: [
         headerParamOverride: [
@@ -141,6 +161,10 @@ defmodule ApicalTest.Parameters.HeaderTest do
     def x_custom("error_list"), do: {:error, message: "list"}
     def x_custom(_, true), do: {:ok, "explode"}
     def x_custom(_, level), do: {:ok, level}
+
+    def defined_marshalling("true", :atom), do: {:ok, true}
+    def defined_marshalling("47", :atom), do: {:ok, 47}
+    def defined_marshalling(_, :atom), do: {:error, "invalid"}
   end
 
   use ApicalTest.EndpointCase
@@ -535,6 +559,57 @@ defmodule ApicalTest.Parameters.HeaderTest do
                conn
                |> Conn.put_req_header("style-custom-override", "ok")
                |> get("/override/")
+               |> json_response(200)
+    end
+  end
+
+  describe "for a marshall-defined parameter" do
+    test "works with a valid value", %{conn: conn} do
+      assert %{"marshal-defined" => true} =
+               conn
+               |> Conn.put_req_header("marshal-defined", "true")
+               |> get("/optional/")
+               |> json_response(200)
+
+      assert %{"marshal-defined" => 47} =
+               conn
+               |> Conn.put_req_header("marshal-defined", "47")
+               |> get("/optional/")
+               |> json_response(200)
+    end
+
+    test "400 with an invalid value", %{conn: conn} do
+      assert_raise Apical.Exceptions.ParameterError,
+                   "Parameter Error in operation headerParamOptional (in header): invalid",
+                   fn ->
+                     conn
+                     |> Conn.put_req_header("marshal-defined", "invalid")
+                     |> get("/optional/")
+                   end
+    end
+  end
+
+  describe "disabled validations" do
+    test "happens with validate: false", %{conn: conn} do
+      # note that this is still marshalled
+      assert %{"validate-disabled" => true} =
+               conn
+               |> Conn.put_req_header("validate-disabled", "true")
+               |> get("/optional/")
+               |> json_response(200)
+
+      assert %{"validate-disabled" => "invalid"} =
+               conn
+               |> Conn.put_req_header("validate-disabled", "invalid")
+               |> get("/optional/")
+               |> json_response(200)
+    end
+
+    test "happens with marshal: false", %{conn: conn} do
+      assert %{"marshal-disabled" => "47"} =
+               conn
+               |> Conn.put_req_header("marshal-disabled", "47")
+               |> get("/optional/")
                |> json_response(200)
     end
   end