+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Apical Changelog
 
-## 0.1
+## 0.1.0
 
 ### initial release
 
@@ -22,6 +22,15 @@
   - application/form-encoded
 - options targettable by tag or operation_id
 - supports internal $refs
+
+## 0.2.0
+
+- support for grouping operationId by "group"
+- support for using Plug only, with no Phoenix dependency
+- support for aliasing operationId to functions
+- support for custom parameter marshalling functions
+- support for disabling marshalling or validation
+- support for using Apical to test requests against OpenAPI schemas
 
 ## Future planned support